--- src/components/Guide/constants.js
@@ -1,1 +1,0 @@
-export const MAX_WIDTH = 1200;
