+++ src/constants.js
@@ -1,1 +1,2 @@
 export const COOKIE_LANGUAGE_KEY = 'lng';
+export const MAX_WIDTH = 1200;
